requirements.txt
matplotlib==3.7.2
numpy==1.25.1
Pillow==10.0.0
seaborn==0.12.2
torch==1.9.1+cu111
transformers==4.30.2
